requirements.txt
astunparse==1.6.3
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
glfw==2.2.0
grpcio==1.63.0
huggingface-hub==0.23.0
imageio-ffmpeg==0.4.3
imgui==1.3.0
itsdangerous==2.2.0
joblib==1.4.2
keras==3.3.3
libclang==18.1.1
markdown-it-py==3.0.0
mdurl==0.1.2
ml-dtypes==0.3.2
moviepy==1.0.3
namex==0.0.8
nvidia-cuda-runtime-cu12==12.4.127
nvidia-pyindex==1.0.9
optree==0.11.0
pandas==1.5.3
pillow==10.3.0
proglog==0.1.10
protobuf==4.25.3
pyasn1-modules==0.2.8
PyOpenGL==3.1.5
pyperlin==1.1.0
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyspng==0.1.1
regex==2024.5.10
requests-oauthlib==1.3.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.4.2
scikit-video==1.1.11
seaborn==0.13.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.8.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.37.0
tensorrt==10.0.1
tensorrt-cu12==10.0.1
tensorrt-cu12-bindings==10.0.1
tensorrt-cu12-libs==10.0.1
threadpoolctl==3.5.0
tokenizers==0.15.2
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
transformers==4.36.2